+++ Changelog.md
@@ -7,6 +7,7 @@
 - Removed 'HDRCore-ToneMapping' because of use of DGTonemap (Dendraspis, JJKylee, 44vince44, [#369](/../../issues/369))
 - aomenc options optimized (Dendraspis, Neltulz, BlueSwordM , [#361](/../../issues/361))
 - Fix rightclick on aomenc options doesn't show help (Dendraspis, [#365](/../../issues/365))
+- Expand 'Restore Profiles' to preserve possible custom profiles (Dendraspis)
 
 
 2.1.5.2 Beta (2020-12-15)